+++ NEWS.md
@@ -1,6 +1,7 @@
 # FSA 0.4.31 ongoing
 * `addZeroCatch()`: Modified.  Slight modifications to help file.  Fixed bug related to error checking the number of variables.  Added some tests.
 * `perc()`: Added.
+* `psdVal()`: Modified.  Changed name for the "zero" group to "substock."
 * `swvCode()`: Modified.  Removed ability to Stangle the code and thus removed `method=`.  Modified code to allow usage of .Rmd files in addition to .Rnw files.